refactor(auth): share jwt utilities
